dev -> master

+++ src/action/action.controller.ts
@@ -6,13 +6,16 @@ import {
   Param,
   Patch,
   Post,
+  Session,
   UseGuards,
 } from '@nestjs/common';
 import { ActionService } from './action.service';
 import { CreateActionDto } from './dto/create-action.dto';
 import { UpdateActionDto } from './dto/update-action.dto';
-import { SessionUserId } from '../auth/decorators/session-user.decorator';
 import { SessionGuard } from '../auth/guards/session.guard';
+import { AuthSession } from '../auth/auth.types';
+import { Pagination } from '../shared/pagination/pagination.decorator';
+import { PaginationParams } from '../shared/pagination/paginataion.types';
 
 @UseGuards(SessionGuard)
 @Controller('action')
@@ -20,31 +23,43 @@ export class ActionController {
   constructor(private readonly actionService: ActionService) {}
 
   @Post()
-  create(@Body() dto: CreateActionDto, @SessionUserId() userId: string) {
+  create(
+    @Body() dto: CreateActionDto,
+    @Session() { user: { id: userId } }: AuthSession,
+  ) {
     return this.actionService.create({ userId, dto });
   }
 
   @Get()
-  findAll(@SessionUserId() userId: string) {
-    return this.actionService.findAll({ userId });
+  findAll(
+    @Session() { user: { id: userId, role } }: AuthSession,
+    @Pagination() filter: PaginationParams,
+  ) {
+    return this.actionService.findAll({ userId, role, filter });
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string, @SessionUserId() userId: string) {
-    return this.actionService.findOne({ id, userId });
+  findOne(
+    @Param('id') id: string,
+    @Session() { user: { id: userId, role } }: AuthSession,
+  ) {
+    return this.actionService.findOne({ id, userId, role });
   }
 
   @Patch(':id')
   update(
     @Param('id') id: string,
     @Body() dto: UpdateActionDto,
-    @SessionUserId() userId: string,
+    @Session() { user: { id: userId, role } }: AuthSession,
   ) {
-    return this.actionService.update({ id, userId, dto });
+    return this.actionService.update({ id, userId, role, dto });
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string, @SessionUserId() userId: string) {
-    return this.actionService.remove({ id, userId });
+  remove(
+    @Param('id') id: string,
+    @Session() { user: { id: userId, role } }: AuthSession,
+  ) {
+    return this.actionService.remove({ id, userId, role });
   }
 }
