+++ src/source/source.controller.ts
@@ -6,13 +6,16 @@ import {
   Param,
   Patch,
   Post,
+  Session,
   UseGuards,
 } from '@nestjs/common';
 import { SourceService } from './source.service';
 import { CreateSourceDto } from './dto/create-source.dto';
 import { UpdateSourceDto } from './dto/update-source.dto';
-import { SessionUserId } from '../auth/decorators/session-user.decorator';
 import { SessionGuard } from '../auth/guards/session.guard';
+import { AuthSession } from '../auth/auth.types';
+import { Pagination } from '../shared/pagination/pagination.decorator';
+import { PaginationParams } from '../shared/pagination/paginataion.types';
 
 @UseGuards(SessionGuard)
 @Controller('source')
@@ -20,31 +23,43 @@ export class SourceController {
   constructor(private readonly sourceService: SourceService) {}
 
   @Post()
-  create(@Body() dto: CreateSourceDto, @SessionUserId() userId: string) {
+  create(
+    @Body() dto: CreateSourceDto,
+    @Session() { user: { id: userId } }: AuthSession,
+  ) {
     return this.sourceService.create({ userId, dto });
   }
 
   @Get()
-  findAll(@SessionUserId() userId: string) {
-    return this.sourceService.findAll({ userId });
+  findAll(
+    @Session() { user: { id: userId, role } }: AuthSession,
+    @Pagination() filter: PaginationParams,
+  ) {
+    return this.sourceService.findAll({ userId, role, filter });
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string, @SessionUserId() userId: string) {
-    return this.sourceService.findOne({ id, userId });
+  findOne(
+    @Param('id') id: string,
+    @Session() { user: { id: userId, role } }: AuthSession,
+  ) {
+    return this.sourceService.findOne({ id, userId, role });
   }
 
   @Patch(':id')
   update(
     @Param('id') id: string,
     @Body() dto: UpdateSourceDto,
-    @SessionUserId() userId: string,
+    @Session() { user: { id: userId, role } }: AuthSession,
   ) {
-    return this.sourceService.update({ id, userId, dto });
+    return this.sourceService.update({ id, userId, role, dto });
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string, @SessionUserId() userId: string) {
-    return this.sourceService.remove({ id, userId });
+  remove(
+    @Param('id') id: string,
+    @Session() { user: { id: userId, role } }: AuthSession,
+  ) {
+    return this.sourceService.remove({ id, userId, role });
   }
 }
